requirements.txt
# Core CLI and Framework
typer[all]==0.9.0
rich==13.7.0
click==8.1.7

# Local LLM Integration
ollama==0.1.7
requests==2.31.0

# Vector Store and Embeddings
chromadb==0.4.15
sentence-transformers==2.2.2
numpy==1.24.3

# Configuration and Environment
pyyaml==6.0.1
python-dotenv==1.0.0
pydantic==2.5.0

# API and Plugin Support
httpx==0.25.2
aiohttp==3.9.0
openapi-spec-validator==0.7.1
dnspython==2.4.2
python-whois==0.8.0
feedparser==6.0.10
urllib3==2.0.7

# Security and Cryptography
cryptography==41.0.7
bcrypt==4.1.2

# Data Processing
pandas==2.1.4
jsonschema==4.20.0

# Logging and Monitoring
loguru==0.7.2

# Optional Cloud Providers (opt-in)
openai==1.3.0
anthropic==0.7.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.0

transformers==4.26.1
huggingface_hub==0.14.1
tokenizers==0.13.3
